Move WPBottomSheetDialogFragment out of the login library
